requirements.txt
google-api-python-client>=2.111.0
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
openai>=1.2.4
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
pytest>=9.0.1

